Add button for uploading file to form, restructure component
