+++ viewer/v2client/src/components/create/file-adder.vue
@@ -1,0 +1,186 @@
+<script>
+import * as _ from 'lodash';
+import * as VocabUtil from '@/utils/vocab';
+import * as RecordUtil from '@/utils/record';
+import * as StringUtil from '@/utils/string';
+
+export default {
+  name: 'file-adder',
+  data () {
+    return {
+      userIsDropping: false,
+      invalidFile: false,
+      droppedFile: {},
+    }
+  },
+  methods: {
+    outputData(data) {
+      this.$emit('output', data);
+    },
+    openPicker() {
+      this.$refs.FilePicker.click();
+    },
+    initFilePicker() {
+      const self = this;
+      this.$refs.FilePicker.addEventListener('input', function(e) {
+        const reader = new FileReader();
+        reader.onloadend = function() {
+          try {
+            const data = JSON.parse(this.result);
+            self.droppedFile = data;
+          } catch (e) {
+            self.invalidFile = true;
+          }
+        };
+        reader.readAsText(e.target.files[0]);
+      });
+    },
+    initDropzone() {
+      window.addEventListener("dragover",function(e){
+        e = e || event;
+        e.preventDefault();
+      }, false);
+      window.addEventListener("drop",function(e){
+        e = e || event;
+        e.preventDefault();
+      }, false);
+      const self = this;
+      this.$refs.dropzone.addEventListener('dragenter', function(e) {
+        e.preventDefault();
+        self.userIsDropping = true;
+      });
+      this.$refs.dropzone.addEventListener('dragleave', function(e) {
+        e.preventDefault();
+        self.userIsDropping = false;
+      });
+      this.$refs.dropzone.addEventListener('drop', function(e) {
+        e.preventDefault();
+        const reader = new FileReader();
+        reader.onloadend = function() {
+          try {
+            const data = JSON.parse(this.result);
+            self.droppedFile = data;
+          } catch (e) {
+            self.invalidFile = true;
+          }
+        };
+        reader.readAsText(e.dataTransfer.files[0]);
+      });
+    },
+  },
+  events: {
+  },
+  computed: {
+    user() {
+      return this.$store.getters.user;
+    },
+    settings() {
+      return this.$store.getters.settings;
+    },
+  },
+  components: {
+  },
+  watch: {
+    'droppedFile': function(val) {
+      if (val.hasOwnProperty('@graph')) {
+        const inspectorObject = RecordUtil.splitJson(val);
+        this.outputData(RecordUtil.prepareDuplicateFor(inspectorObject, this.user, this.settings));
+      } else {
+        this.invalidFile = true;
+      }
+    },
+    'invalidFile': function(val) {
+      if (val === true) {
+        setTimeout(() => {
+          this.invalidFile = false;
+          this.userIsDropping = false;
+        }, 3000);
+      }
+    },
+  },
+  mounted() { // Ready method is deprecated in 2.0, switch to "mounted"
+    this.$nextTick(() => {
+      this.initDropzone();
+      this.initFilePicker();
+    });
+  },
+};
+</script>
+
+<template>
+  <div class="FileAdder">
+    <p>Notera att denna funktion automatiskt rensar och byter ut vissa värden som inte ska tas med till den nya posten. Till detta hör bland annat en posts ID, information om vem som skapat posten och när den skapades.</p>
+    <button class="btn btn-primary" @click="openPicker">{{ 'Choose file' | translatePhrase }}</button>
+    <input type="file" class="FilePicker" ref="FilePicker" accept=".jsonld,application/ld+json,text/*" />
+    <hr/>{{ 'or' | translatePhrase }}<hr/>
+    <div class="Dropzone" :class="{'is-active': userIsDropping, 'is-invalid': invalidFile}">
+      <div class="Dropzone-mask" ref="dropzone"></div>
+      <div class="Dropzone-container">
+        <div class="Dropzone-description" v-show="!invalidFile">{{'Drop your file here' | translatePhrase}}</div>
+        <div class="Dropzone-description" v-show="invalidFile">{{'Invalid file' | translatePhrase}}</div>
+      </div>
+    </div>
+  </div>
+</template>
+
+<style lang="less">
+
+.FileAdder {
+  border: 1px solid @form-border;
+  background-color: @white;
+  display: flex;
+  padding: 1em;
+  flex-direction: column;
+  align-items: center;
+  p {
+    margin-bottom: 2em;
+  }
+  .FilePicker {
+    width: 1px;
+    height: 1px;
+    opacity: 0;
+  }
+  hr {
+    display: inline-block;
+    width: 25%;
+  }
+  .Dropzone {
+    border: 3px solid @gray;
+    border-radius: 1em;
+    height: 25vh;
+    width: 80%;
+    &.is-active {
+      background-color: fadeout(@brand-primary, 50%);
+      border: 3px dashed @brand-primary;
+    }
+    &.is-invalid {
+      background-color: fadeout(@brand-danger, 50%);
+      border: 3px dashed @brand-danger;
+    }
+    &-title {
+      font-size: 24px;
+      font-size: 2.4rem;
+      font-weight: 500;
+    }
+    &-container {
+      display: flex;
+      position: relative;
+      top: -102%;
+      flex-direction: row;
+      align-items: center;
+      height: inherit;
+    }
+    &-description {
+      text-align: center;
+      width: 100%;
+    }
+    &-mask {
+      position: relative;
+      width: inherit;
+      height: inherit;
+      z-index: 999;
+      background-color: rgba(0, 0, 0, 0);
+    }
+  }
+}
+</style>
