+++ viewer/v2client/src/resources/json/i18n.json
@@ -49,6 +49,7 @@
     "Loading documents" : "Laddar dokument",
     "Choose": "Välj",
     "Chosen": "Vald",
+    "Choose file": "Välj fil",
     "Help": "Hjälp",
     "Loading editor" : "Laddar redigeringen",
     "View MARC21" : "Granska MARC21",
@@ -117,6 +118,7 @@
     "If no matching linked entity is found you can create and link. This will create a linked entity containing the information in the entity chosen for linking": "Om du inte lyckas identifiera en matchande länkad entitet kan du skapa och länka. Detta kommer att skapa en länkad entitet med samma innehåll som entiteten som valts för länkning",
     "Filter by": "Filtrera",
     "Showing": "Visar",
+    "or": "eller",
     "Tools": "Verktyg",
     "Show as": "Visa som",
     "Expand all": "Expandera alla",
